MAINT: Refactor mapping functions
