+++ frontend/src/components/project/mapping/MappingData.ts
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+import type { MappingData } from "./types";
+
+export const MappingDataContext = createContext<MappingData | null>(null);
